+++ mobile/android/geckoview/src/androidTest/java/org/mozilla/geckoview/test/PanZoomControllerTest.kt
@@ -251,9 +251,8 @@ class PanZoomControllerTest : BaseSessionTest() {
         val down = MotionEvent.obtain(
                 downTime, SystemClock.uptimeMillis(), MotionEvent.ACTION_DOWN, x, y, 0);
 
-        @Suppress("DEPRECATION")
-        val result = mainSession.panZoomController.onTouchEventForResult(down)
-
+        val result = mainSession.panZoomController.onTouchEventForDetailResult(down)
+                .map { value -> value!!.handledResult() }
         val up = MotionEvent.obtain(
                 downTime, SystemClock.uptimeMillis(), MotionEvent.ACTION_UP, x, y, 0);
 
@@ -457,8 +456,8 @@ class PanZoomControllerTest : BaseSessionTest() {
         val down = MotionEvent.obtain(
                 downTime, SystemClock.uptimeMillis(), MotionEvent.ACTION_DOWN, 50f, 90f, 0)
 
-        @Suppress("DEPRECATION")
-        val result = mainSession.panZoomController.onTouchEventForResult(down)
+        val result = mainSession.panZoomController.onTouchEventForDetailResult(down)
+                .map { value -> value!!.handledResult() }
         var move = MotionEvent.obtain(
                 downTime, SystemClock.uptimeMillis(), MotionEvent.ACTION_MOVE, 50f, 70f, 0)
         mainSession.panZoomController.onTouchEvent(move)